requirements.txt
aenum==2.1.2
appnope==0.1.0
astroid==2.1.0
backcall==0.1.0
cffi==1.11.5
Click==7.0
colorful==0.5.0
decorator==4.3.2
dicttoxml==1.7.4
greenlet==0.4.13
ipython==7.2.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.2
lazy-object-proxy==1.3.1
lxml==4.3.1
mccabe==0.6.1
parso==0.3.3
pexpect==4.6.0
pickleshare==0.7.5
prettyprinter==0.14.0
prompt-toolkit==2.0.8
ptyprocess==0.6.0
Pygments==2.3.1
pylint==2.2.2
pyone==5.7.85
python-dotenv==0.10.1
readline==6.2.4.1
six==1.12.0
tblib==1.3.2
traitlets==4.3.2
wcwidth==0.1.7
wrapt==1.11.1
xmltodict==0.12.0
